Mobile: add react-cosmos
