+++ apps/tlon-mobile/.gitignore
@@ -1,6 +1,7 @@
 # Generated
 tailwind.css
 tailwind.json
+cosmos.imports.ts
 
 # Created by https://www.toptal.com/developers/gitignore/api/react,reactnative,macos,visualstudiocode
 # Edit at https://www.toptal.com/developers/gitignore?templates=react,reactnative,macos,visualstudiocode
