+++ apps/tlon-mobile/package.json
@@ -19,7 +19,8 @@
     "test": "vitest --watch false",
     "tsc": "tsc --noEmit",
     "postinstall": "npm run generate",
-    "eas-build-post-install": "./post-install"
+    "eas-build-post-install": "./post-install",
+    "cosmos": "cosmos-native"
   },
   "lint-staged": {
     "*.{js,jsx,ts,tsx,json,md,html}": [
@@ -108,6 +109,8 @@
     "postcss": "^8.4.26",
     "prettier": "^3.2.5",
     "prettier-plugin-tailwindcss": "^0.5.4",
+    "react-cosmos": "^6.1.1",
+    "react-cosmos-native": "^6.1.1",
     "react-native-svg-transformer": "^1.3.0",
     "tailwindcss": "^3.3.3",
     "typescript": "^5.1.3",
