+++ apps/tlon-mobile/src/App.fixture.tsx
@@ -1,0 +1,4 @@
+import App from "./App.main";
+
+// eslint-disable-next-line
+export default () => <App />;
