+++ apps/tlon-mobile/src/cosmos.decorator.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { SafeAreaView } from "react-native";
+import { SafeAreaProvider } from "react-native-safe-area-context";
+
+// eslint-disable-next-line
+export default ({ children }: { children: React.ReactNode }) => (
+  <SafeAreaProvider>
+    <SafeAreaView>{children}</SafeAreaView>
+  </SafeAreaProvider>
+);
