+++ apps/tlon-mobile/src/fixtures/Button.fixture.tsx
@@ -1,0 +1,9 @@
+import { Button } from "@tloncorp/ui";
+
+export default {
+  primary: () => (
+    <Button>
+      <Button.Text>Primary</Button.Text>
+    </Button>
+  ),
+};
