+++ apps/tlon-mobile/src/fixtures/HeaderButton.fixture.tsx
@@ -1,0 +1,7 @@
+import { HeaderButton } from "../components/HeaderButton";
+
+export default {
+  primary: () => (
+    <HeaderButton title="Button" onPress={() => console.log("press")} />
+  ),
+};
