+++ apps/tlon-mobile/src/fixtures/TlonButton.fixture.tsx
@@ -1,0 +1,7 @@
+import { TlonButton } from "../components/TlonButton";
+
+export default {
+  primary: () => <TlonButton title="Button" />,
+  secondary: () => <TlonButton title="Button" variant="secondary" />,
+  minimal: () => <TlonButton title="Button" variant="minimal" />,
+};
